后台：步骤审批视图删除功能优化

+++ app/Http/Controllers/Api/Admin/StepApproverController.php
@@ -94,7 +94,17 @@ class StepApproverController extends Controller
      */
     public function destroy($id)
     {
-        StepApprover::destroy($id);
+        $data = StepApprover::with(['departments','steps'])->find($id);
+
+        if($data->departments){
+            abort(400,'该审批视图下有所属部门，不能删除');
+        }
+        if($data->steps){
+            $flowIds = $data->steps->pluck('flow_id')->all();
+            $stepName = $data->steps->pluck('name')->all();
+            abort(400,'流程ID为 ('.implode(',',$flowIds).') 下面步骤名 ('.implode(',',$stepName).')  使用了该审批视图');
+        }
+        $data->delete();
         return $this->response->delete();
     }
 }
